pyproject.toml
[project]
name = "imposter"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.6.4",
    "google-genai>=1.56.0",
    "python-dotenv>=1.2.1",
]
